@@ -35,8 +35,8 @@
     "react-dom": "16.8.6",
     "react-svg-loader": "3.0.3",
     "sinon": "7.5.0",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^4.10.0"
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4"
   },
   "files": []
 }
