@@ -13,13 +13,14 @@
     "@babel/register": "7.9.0",
     "@types/jsdom": "16.2.13",
     "@types/mock-require": "2.0.0",
-    "@types/webpack": "4.41.0",
+    "@types/webpack": "^5.28.1",
     "@types/webpack-dev-server": "^4.0.0",
     "ansi-escapes": "4.3.1",
     "arraybuffer-loader": "1.0.8",
     "autoprefixer": "9.7.4",
     "babel-loader": "8.3.0",
     "browser-resolve": "2.0.0",
+    "buffer": "6.0.3",
     "chai": "4.2.0",
     "chalk": "2.4.2",
     "clean-webpack-plugin": "4.0.0",
@@ -32,14 +33,19 @@
     "mini-css-extract-plugin": "1.6.2",
     "mock-require": "3.0.3",
     "node-sass-glob-importer": "5.3.3",
+    "os-browserify": "0.3.0",
+    "path-browserify": "1.0.1",
     "postcss": "^8.4.27",
     "postcss-loader": "4.3.0",
+    "process": "0.11.10",
     "resolve-url-loader": "5.0.0",
     "sass": "1.64.1",
     "sass-loader": "10.4.1",
     "sinon-chai": "3.3.0",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^4.10.0",
+    "stream-browserify": "3.0.0",
+    "url": "0.11.1",
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4",
     "webpack-livereload-plugin": "3.0.2"
   }
 }
